Moved data to index.jsx, some CSS touchups
